Add SQL testing support and DB connections
This commit adds the ability to connect to a database
by trying first to connect to faure using the environmental
variables CSU_NAME and CSU_ID, then with the system properties
username and password.  System properties can be set by editing
the run script and adding -Dusername=x -Dpassword=y, similar to
configuring the SOCKS proxy. If faure can't be connected to then
it falls back to a local database for the purpose of testing with
travis.

If the database still can't be reached it configures a SOCKS proxy
to localhost port 9008(because we're team 8) and repeats the process.

The first ~ 100 entries are uploaded to travis for our unit tests.
Unit tests should not rely on the specific entries that are in the
database or they might be incompatible with one of the databases.

Closes #207, closes #210, closes #214